rename ContributorCategorySelect -> ContributorsStatusSelect
