--- src/app/(user)/contributors/_components/ContributorCategorySelect.tsx
@@ -1,21 +1,0 @@
-import ContributorCategorySelectItem from './ContributorCategorySelectItem'
-import styles from './ContributorCategorySelect.module.scss'
-
-export default function ContributorCategorySelect() {
-  return (
-    <div className={styles.container}>
-      <p className={styles.label}>Quick Filters</p>
-      <ul className={styles.content}>
-        <ContributorCategorySelectItem value="none">
-          None
-        </ContributorCategorySelectItem>
-        <ContributorCategorySelectItem value="low_on_tasks">
-          Low on tasks
-        </ContributorCategorySelectItem>
-        <ContributorCategorySelectItem value="no_tasks">
-          No task assigned
-        </ContributorCategorySelectItem>
-      </ul>
-    </div>
-  )
-}
